Fix post render bug
