+++ src/Components/wall.js
@@ -1,9 +1,10 @@
 import { signOut } from 'firebase/auth';
+import { query, collection, getDocs, onSnapshot } from 'firebase/firestore'
 import {
   addComment,
-  showComments,
   auth,
   deleteComment,
+  db,
 } from '../lib/firebase';
 
 // eslint-disable-next-line no-unused-vars
@@ -38,8 +39,36 @@ export const Muro = (onNavigate) => {
   muroDiv.appendChild(newCommentBox);
   muroDiv.appendChild(signOutButton);
   newCommentBox.appendChild(btnDelete);
-  // eslint-disable-next-line no-console
-  console.log(muroDiv);
+
+  const q = query(collection(db, 'comment'));
+  getDocs(q).then((querySnapshot) => {
+    querySnapshot.forEach((docum) => {
+      const comment = docum.data();
+      newCommentBox.innerHTML += `
+      <div class="comment">
+      <h5> ${comment.author}</h5>
+      <p> ${comment.text}</p>
+      <button>Eliminar</button>
+      </div>
+      `;
+    });
+  });
+
+  onSnapshot(q, (querySnapshot) => {
+    let newComment = '';
+    querySnapshot.forEach((docum) => {
+      const comment = docum.data();
+      newComment += `
+      <div class="comment">
+      <h5> ${comment.author}</h5>
+      <p> ${comment.text}</p>
+      <button>Eliminar</button>
+      </div>
+      `;
+    });
+    document.getElementById('postsPrints').innerHTML = newComment;
+  });
+
   postButton.addEventListener('click', (event) => {
     event.preventDefault();
     addComment(commentBox.value, auth.currentUser.email);
@@ -56,8 +85,6 @@ export const Muro = (onNavigate) => {
     });
   });
   // eslint-disable-next-line no-console
-  console.log(showComments());
-console.log(newCommentBox);
   newCommentBox.querySelectorAll('.btnDelete').forEach((button) => {
     console.log(button);
     button.addEventListener('click', (e) => {
